+++ packages/fs-detectors/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @vercel/fs-detectors
 
+## 5.11.0
+
+### Minor Changes
+
+- [services] add support for background workers to vc dev ([#15434](https://github.com/vercel/vercel/pull/15434))
+
+### Patch Changes
+
+- [services] exclude service route prefixes on subdomain rewrites ([#15508](https://github.com/vercel/vercel/pull/15508))
+
 ## 5.10.3
 
 ### Patch Changes