+++ packages/koa/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @vercel/koa
 
+## 0.1.32
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/node@5.6.16
+
 ## 0.1.31
 
 ### Patch Changes